Changed Model Data Type
Changed the data type from double to decimal to fix the error where it
showed up to 6 after the decimal numbers.

+++ NFLData/NFLData/Models/PlayerModel.cs
@@ -28,7 +28,7 @@ namespace NFLData.Models
 
         [DefaultValue(0)]
         [DisplayName("Yards Per Attempt")]
-        public double YardsPerAttempt { get; set; }
+        public decimal YardsPerAttempt { get; set; }
 
         [DefaultValue(0)]
         [DisplayName("Touchdowns")]
@@ -52,7 +52,7 @@ namespace NFLData.Models
 
         [DefaultValue(0)]
         [DisplayName("Rating")]
-        public double Rate { get; set; }
+        public decimal Rate { get; set; }
 
         [DefaultValue(0)]
         [DisplayName("Total Points")]
